USB control of specific servos
It is possible to control a servo by index now in addition to by servo object.

Servo objects renamed to servox, where x is 0, 1, 2, ... 15

All servo objects are initialized when the servo driver is initialized. Initializing a servo yourself should supersede the assigned servo. This allows you to define your own servo (with nice names) and add them to the servo driver.

USB vendor requests updated. One can now specify a servo index (0-15) and a position, and the servo at the index will be set to the specified position.